Done with graph dependencies
