+++ pyproject.toml
@@ -1,3 +1,16 @@
+[tool.poetry]
+name = "anomaly-detection-project"
+version = "0.0.1"
+description = "Anomaly Detection Bootcamp Project"
+authors = ["Vector AI Engineering <vectorinstitute.ai>"]
+license = "MIT License"
+readme = "README.md"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "License :: OSI Approved :: MIT License",
+    "Operating System :: OS Independent",
+]
+
 [build-system]
 requires = ["setuptools", "wheel"]
 build-backend = "setuptools.build_meta"
@@ -9,3 +22,17 @@ use_parentheses = true
 
 [tool.black]
 line-length = 79
+
+[tool.poetry.dependencies]
+python = ">=3.8,<3.9"
+numpy = "^1.21.6"
+pandas = "^1.3.5"
+tqdm = "^4.64.1"
+torch = "^1.13.0+cu116"
+scikit-learn = "^1.0.2"
+delu = "^0.0.13"
+torch-geometric = "^2.3.1"
+scipy = "^1.7.3"
+rtdl = "^0.0.13"
+xgboost = "^1.6.2"
+pyod = "^1.0.9"
