+++ video_anomaly_detection/UCSD_pedestrian/DMAD/model_config.yaml
