+++ video_anomaly_detection/UCSD_pedestrian/DMAD/run.py
