+++ video_anomaly_detection/UCSD_pedestrian/Dataset/data_config.yaml
