+++ video_anomaly_detection/UCSD_pedestrian/Demo_run.py
