+++ video_anomaly_detection/UCSD_pedestrian/README.md
