+++ video_anomaly_detection/UCSD_pedestrian/images/diversity.png
