+++ video_anomaly_detection/UCSD_pedestrian/images/eq1.png
