+++ video_anomaly_detection/UCSD_pedestrian/images/eq2.png
