+++ video_anomaly_detection/UCSD_pedestrian/images/eq3.png
