+++ video_anomaly_detection/UCSD_pedestrian/images/eq4.png
