+++ video_anomaly_detection/UCSD_pedestrian/images/eq5.png
