+++ video_anomaly_detection/UCSD_pedestrian/images/eq6.png
