+++ video_anomaly_detection/UCSD_pedestrian/images/eq7.png
