+++ video_anomaly_detection/UCSD_pedestrian/images/main_model.png
