+++ video_anomaly_detection/UCSD_pedestrian/images/results.png
